Rename VhostUserSlave->VhostUserMasterReqHandler
Our naming is getting confusing because we've essentially
got two things called 'slave' - one is the program that's not
the VMM that reads commands sent by the VMM, and the other is the
fd/channel which commands from the slave are sent back to the VMM.

In 34bec2ceb63 Daniel introduced VhostUserSlaveReqHandler as the
trait to handle requests *from* the slave sent over the slave
channel.
Lets follow the same naming scheme and make 'VhostUserSlave'
become VhostUserMasterReqHandler.

So now we have:

   VhostUserSlaveReqHandler   - Handles requests from the slave
   VhostUserMasterReqHandler  - Handles requests from the master/VMM

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>